backend-service/requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.0.0
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
firebase-admin==6.1.0
google-generativeai==0.3.1
PyGithub==2.1.1
pytest==7.4.0
tqdm==4.66.1
flask-cors==4.0.0